fix(compiler): simplify unary Int32 conversions (#61)
* fix(compiler): simplify unary Int32 conversions

Haxe Int32 negation and bitwise complement already produce native Go int32 values after their operand is narrowed. Widen those results directly to the Go int carrier instead of wrapping them in a second int32 conversion.

Extend the arithmetic snapshot with both unary operations and protect their generated shape in the lint contract. The focused runtime snapshot and the complete changed-surface suite pass.

* test: refresh unary Int32 snapshots

The unary lowering change affects shared Haxe standard-library paths in nine additional snapshot cases. Refresh those generated expectations so the full 332-case CI sweep checks the new direct widening shape everywhere it appears.

All nine affected snapshots pass with the locked Haxe/Lix shim, including runtime compilation and the slow Int64 fixture.

* test: refresh incident API unary output

Refresh the portable and metal incident API snapshots after unary Int lowering stopped emitting a redundant second int32 conversion.\n\nThis keeps the checked-in example output aligned with the compiler while preserving the same runtime behavior. Both incident API profiles and the generated lint contract pass locally.

+++ test/snapshot/core/static_dynamic_optional_primitive/intended/main.go
@@ -139,7 +139,7 @@ func main() {
 		if enabled.(bool) {
 			hx_if_10 = count.(int)
 		} else {
-			hx_if_10 = int(int32(-int32(count.(int))))
+			hx_if_10 = int(-int32(count.(int)))
 		}
 		return hx_if_10
 	}
@@ -168,7 +168,7 @@ func main() {
 		if enabled.(bool) {
 			hx_if_12 = count.(int)
 		} else {
-			hx_if_12 = int(int32(-int32(count.(int))))
+			hx_if_12 = int(-int32(count.(int)))
 		}
 		return hx_if_12
 	}
@@ -185,7 +185,7 @@ func main() {
 		if enabled.(bool) {
 			hx_if_13 = count.(int)
 		} else {
-			hx_if_13 = int(int32(-int32(count.(int))))
+			hx_if_13 = int(-int32(count.(int)))
 		}
 		return hx_if_13
 	}
